edit: 역할에 따라 클래스 분리하기
