+++ src/main/java/baseball/Game.java
@@ -1,0 +1,11 @@
+package baseball;
+
+public class Game {
+
+    public void startGame() {
+        Computer computer = new Computer();
+        User user = new User();
+        
+    }
+
+}
